// src/types/types.ts
export interface searchData {
  place_name: string; // 상가명
  category_name: string; // 카테고리
  address_name: string; // 주소
  road_address_name: string; // 도로명 주소
  phone: string; // 전화번호
  searchCnt: number; // 조회수
  reviewCnt: number; // 리뷰수
  like: boolean; // 즐겨찾기 여부
}